worker: move the guardian init to a separate file
- also, rename gdn to guardian
